requirements.txt
appdirs==1.4.4
asgiref==3.3.1
astroid==2.4.2
backcall==0.2.0
black==20.8b1
click==7.1.2
decorator==4.4.2
Django==3.1.6
django-extensions==3.1.0
djangorestframework==3.12.2
factory-boy==3.2.0
Faker==6.0.0
ipython==7.20.0
ipython-genutils==0.2.0
isort==5.7.0
jedi==0.18.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
mypy-extensions==0.4.3
parso==0.8.1
pathspec==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
pkg-resources==0.0.0
prompt-toolkit==3.0.14
psycopg2==2.8.6
ptyprocess==0.7.0
Pygments==2.7.4
pylint==2.6.0
python-dateutil==2.8.1
pytz==2021.1
regex==2020.11.13
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
traitlets==5.0.5
typed-ast==1.4.2
typing-extensions==3.7.4.3
wcwidth==0.2.5
wrapt==1.12.1
